Removed axioms from Core.v
Axioms like the excluded middle and function extenionality do not
make the theory of Coq inconsistent. However the axiom of the
excluded middle states that every proposition has either a proof
or a refutation. This is definitely not valid because there are
undecidable propositions.

Therefore we try to stay constructive as long as no showstopper
appears.

	modified:   src/Core.v