Added strict 3 way comparison to Relation and Sortable
A normal 3 way comparison contains for LT a proof of 'less_equal'
and for GT a proof of 'greater_equal'.

A strict 3 way comparison contains for LT a proof of 'less_equal' and
'not greater_equal' and for GT a proof of 'greater_equal' and 'not
less_equal'. I.e. equivalent elements are never possible for LT and
GT.

	modified:   src/lib/Binary_tree.v
	modified:   src/lib/Core.v
	modified:   src/lib/Red_black_tree.v