Fix: Insertion and replacement not handled correctly
Now the function inserting an element into a red black tree
gets a replace flag indicating whether an equivalent element
shall be replace by the element to be inserted or that the
element to be inserted will be inserted at the end of all
equivalent elements.

Therefore only one 'compare3' decider is necessary in sortable
elements. 'compare3 a b' returns either 'LT' with a proof of
'a <= b' and '~ b <= a' or 'EQV' with a proof of 'a <= b' and
'b <= a' or 'GT' with a proof of 'b <= a' and '~ a <= b'.

The 'Inserted.R' relation has been augmented by a replace flag
indicating if replacement or insertion shall happen.

	modified:   src/lib/Binary_search_tree.v
	modified:   src/lib/Core.v
	modified:   src/lib/Red_black_tree.v