Added theorem 'accessible_induction'
The theorem states

      (forall x, Acc R x -> (forall y, R y x ->  P y) -> P x)
      ->
      forall x, Acc R x -> P x.

which is better usable than the induction law derived by Coq.

If we want to prove forall x, Acc R x -> Px we just prove the premise
which in addition to Acc R x adds an induction hypothesis. Therefore
the goal P x is easier to prove because the context is richer.

	modified:   src/lib/Core.v